Fixed 404 page
Fixed an issue where the 404 error page would rupture the fabric of space and time and loop endlessly instead of playing the correct animation if viewed from inside a subfolder. This has been an issue for literally as long as the 404 page has existed and it's a literal one character fix. Lol, lmao even