Fixed 404 page (#53)
Fixed an issue where the 404 error page would rupture the fabric of space and time and loop endlessly instead of playing the correct animation if viewed from inside a subfolder. This has been an issue for literally as long as the 404 page has existed and it's a literal one character fix. Lol, lmao even

---------

Co-authored-by: trickmint <71721378+trickmint@users.noreply.github.com>